refactor(hooks): move getForwarderStakeUserData() into preparers util
